Extract player creation event.
